Refactor Buttons class with new Buttons file
